+++ app/pb_kits/playbook/pb_checkbox/docs/_checkbox_error.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import { Checkbox } from '../..'
+
+const CheckboxError = () => {
+  return (
+    <div>
+      <Checkbox
+          error
+          name="default name"
+          text="Checkbox label"
+          value="default value"
+      />
+    </div>
+  )
+}
+
+export default CheckboxError