+++ app/pb_kits/playbook/pb_text_input/docs/_text_input_error.jsx
@@ -1,0 +1,17 @@
+import React from 'react'
+import {
+  TextInput,
+} from '../..'
+
+const TextInputError = () => (
+  <div>
+    <TextInput
+        error="Please enter a valid email address"
+        label="Email Address"
+        placeholder="Enter email address"
+        type="email"
+    />
+  </div>
+)
+
+export default TextInputError