+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.5.0] Unreleased
 
+### Changed
+- Added error state to Text Input kit [#554](https://github.com/powerhome/playbook/pull/554)
+
 ## [3.4.0] 2019-1-9
 
 ### Added